Add colorization to Python syntax and fix bugs
This commit uses the pygments module to colorize the terms in the
Python config files.  The pygments module parses the Python
and applies colors based on the sytactical role of terms.

This commit fixes many bugs:
 * how the syntax trees were searched for equivalent statements
 * deciding which side should be merged next into the output when
   the types of nodes are mismatched
 * how the type of RedBaron node was extracted from the first node in a
   contiguous block
 * how contiguous blocks of nodes are identified
 * how the identifier for a pair of contiguous blocks was selected

A few more interaction messages were moved to the localization
dictionary.

To Do: Change colorization to highlight differences in the text.